Revert "Merge branch 'fix/types' into unsync-anthropic-stop-reason"
This reverts commit 80790e63b6585c31f68af9a1cc280f8944d71137, reversing
changes made to a72cf50c81634db0a8c58e5d76401ff4e9642dd0.
