+++ src/providers/google-vertex-ai/chatComplete.ts
@@ -3,14 +3,11 @@
 
 import { GOOGLE_VERTEX_AI } from '../../globals';
 import {
-  AssistantMessage,
   ContentType,
   Message,
   Params,
   ToolCall,
-  ToolMessage,
 } from '../../types/requestBody';
-import { Message as ResponseMessage } from '../../types/responseBody';
 import {
   AnthropicChatCompleteResponse,
   AnthropicChatCompleteStreamResponse,
@@ -22,7 +19,6 @@ import {
   GoogleMessage,
   GoogleMessageRole,
   GoogleToolConfig,
-  PortkeyGeminiMessage,
   SYSTEM_INSTRUCTION_DISABLED_MODELS,
   transformOpenAIRoleToGoogleRole,
   transformToolChoiceForGemini,
@@ -57,7 +53,7 @@ export const VertexGoogleChatCompleteConfig: ProviderConfig = {
         let lastRole: GoogleMessageRole | undefined;
         const messages: GoogleMessage[] = [];
 
-        params.messages?.forEach((message: PortkeyGeminiMessage) => {
+        params.messages?.forEach((message: Message) => {
           // From gemini-1.5 onwards, systemInstruction is supported
           // Skipping system message and sending it in systemInstruction for gemini 1.5 models
           if (
@@ -303,7 +299,7 @@ interface AnthropicToolResultContentItem {
 
 type AnthropicMessageContentItem = AnthropicToolResultContentItem | ContentType;
 
-interface AnthropicMessage extends Omit<Message, 'content'> {
+interface AnthropicMessage extends Message {
   content?: string | AnthropicMessageContentItem[];
 }
 
@@ -322,9 +318,10 @@ interface AnthropicToolContentItem {
 type AnthropicContentItem = AnthorpicTextContentItem | AnthropicToolContentItem;
 
 const transformAssistantMessageForAnthropic = (
-  msg: AssistantMessage
+  msg: Message
 ): AnthropicMessage => {
   let content: AnthropicContentItem[] = [];
+  const containsToolCalls = msg.tool_calls && msg.tool_calls.length;
 
   if (msg.content && typeof msg.content === 'string') {
     content.push({
@@ -343,7 +340,7 @@ const transformAssistantMessageForAnthropic = (
       });
     }
   }
-  if (msg.tool_calls) {
+  if (containsToolCalls) {
     msg.tool_calls.forEach((toolCall: any) => {
       content.push({
         type: 'tool_use',
@@ -359,9 +356,7 @@ const transformAssistantMessageForAnthropic = (
   };
 };
 
-const transformToolMessageForAnthropic = (
-  msg: ToolMessage
-): AnthropicMessage => {
+const transformToolMessageForAnthropic = (msg: Message): AnthropicMessage => {
   return {
     role: 'user',
     content: [
@@ -607,7 +602,7 @@ export const GoogleChatCompleteResponseTransform: (
       provider: GOOGLE_VERTEX_AI,
       choices:
         response.candidates?.map((generation, index) => {
-          let message: ResponseMessage = { role: 'assistant', content: '' };
+          let message: Message = { role: 'assistant', content: '' };
           if (generation.content.parts[0]?.text) {
             message = {
               role: 'assistant',
@@ -679,7 +674,7 @@ export const GoogleChatCompleteStreamChunkTransform: (
     provider: GOOGLE_VERTEX_AI,
     choices:
       parsedChunk.candidates?.map((generation, index) => {
-        let message: ResponseMessage = { role: 'assistant', content: '' };
+        let message: Message = { role: 'assistant', content: '' };
         if (generation.content.parts[0]?.text) {
           message = {
             role: 'assistant',
